+++ .gitignore
@@ -1,5 +1,15 @@
-.env
-
 # Mac OSX temporary files
 .DS_Store
 **/.DS_Store
+
+# editors: emacs, vscode
+*~
+.vscode/
+
+# logs
+logs/
+
+# python
+*.pyc
+**/__pycache__/
+.env
