+++ README.md
@@ -6,7 +6,7 @@ Queries the API every 5 minutes and reports changes to email.
 Can be configured to filter different kinds of events.
 
 
-### Running
+### Setup
 Create a .env file in this directory like so
 ```text
 gmailUsername   = "email.sender@gmail.com"
@@ -14,12 +14,18 @@ gmailPassword   = "mypassword"
 discordBotToken = "xxxxxxxxxxxxxx"
 emailRecipient  = "email.receiver@gmail.com"
 nodeProviderId  = "abc2d-48fgj-32ab3-2a..."
-
 ```
 
+Install dependencies
+```sh
+$ pip install -r requirements.txt
+```
+
+
+### Running
 Run with python, and stop with keyboardinterrupt (crtl-c)
 ```sh
-> python3 node_monitor.py
+$ python3 -m node_monitor
 [2022-10-08 07:52:37.210304]: Starting Node Monitor...
 [2022-10-08 07:52:37.982442]: -- Fetched New Data
 [2022-10-08 07:52:38.650524]: -- Fetched New Data
