+++ node_monitor/__main__.py
@@ -1,0 +1,5 @@
+from node_monitor.node_monitor import NodeMonitor
+
+if __name__ == "__main__":
+    nodemonitor = NodeMonitor()
+    nodemonitor.runloop()
